+++ vortex-array/src/expr/optimizer/rules/mod.rs
@@ -1,10 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::fmt::Debug;
+use std::sync::Arc;
+
 use vortex_error::VortexResult;
 
 use crate::dtype::DType;
 use crate::expr::BoundExpression;
+use crate::expr::ExpressionId;
 use crate::scalar_fn::ScalarFnVTableExt;
 use crate::scalar_fn::fns::cast::Cast;
 
@@ -25,6 +29,28 @@ pub(crate) use nulls::RemoveRedundantFillNull;
 pub(crate) use structural::GetItemFromPack;
 pub(crate) use structural::MergeToPack;
 pub(crate) use structural::SelectFromPack;
+
+/// Shared reference to a bound-expression rewrite rule.
+pub type BoundExpressionRewriteRuleRef = Arc<dyn BoundExpressionRewriteRule>;
+
+/// An equivalence rewrite for bound expressions with a particular root node implementation.
+///
+/// The optimizer invokes a rule only when the expression's root ID equals
+/// [`Self::expression_id`]. Returning `None` means the rule does not match. A replacement must be
+/// semantically equivalent to the input and have exactly the same dtype, including nullability.
+/// The optimizer verifies the dtype and rejects unchanged replacements.
+pub trait BoundExpressionRewriteRule: Debug + Send + Sync + 'static {
+    /// Returns a diagnostic name for this rule.
+    fn name(&self) -> &'static str {
+        std::any::type_name::<Self>()
+    }
+
+    /// Returns the expression node ID handled by this rule.
+    fn expression_id(&self) -> ExpressionId;
+
+    /// Try to rewrite `expr` to a semantically equivalent bound expression.
+    fn rewrite(&self, expr: &BoundExpression) -> VortexResult<Option<BoundExpression>>;
+}
 
 fn preserve_dtype(replacement: BoundExpression, dtype: &DType) -> VortexResult<BoundExpression> {
     if replacement.dtype() == dtype {
