+++ vortex-array/src/expr/optimizer/rules/nulls.rs
@@ -3,11 +3,11 @@
 
 use vortex_error::VortexResult;
 
+use super::BoundExpressionRewriteRule;
 use super::preserve_dtype;
 use crate::expr::BoundExpression;
 use crate::expr::ExpressionId;
 use crate::expr::bound;
-use crate::expr::optimizer::BoundExpressionRewriteRule;
 use crate::scalar_fn::EmptyOptions;
 use crate::scalar_fn::ScalarFnVTable;
 use crate::scalar_fn::ScalarFnVTableExt;
